Add separate Dockerfile for Cloud Run
This is required because we build the app outside of
the Docker image now and copy the result into the
image.
On Cloud Run the app files are missing because of that.
As a result the build fails.
The workaround is to use the old build behavior for Cloud Run.
It's okay because the build does not take that long anyway (3min)